This PR ensures that the order of the polygons in the mesh returned by `tessellate(pts, VoronoiTessellation)` respects the order of the inputs points `pts`, closing #974.

In the process of modifying the function, I also tried some benchmarks and did some slight modifications to the code to reduce un-necessaries allocation. 

The function now preallocates a vector of Connectivity objects and fill it in the correct order inside a for loop.
One modification that significantly reduced allocations was to use the `connect(idxs, NGon)` function rather than the generic `connect(idxs)` one, as I believe you would always have NGons as output of the voronoi tessellation.

Another small improvements in allocations was to use `ntuple` instead of `Tuple(idxs)` to create the temporary indices tuple as the latter allocates (I guess due to type instability since it can't know the tuple length just with the type if called with a vector as input).

I also added a test to verify that each point in the input is `inside` the respective polygon based on index.
Unfortunately I realized that probably due to floating point precision, some of the points would not fall inside the respective polygons. In all these cases, the culprit point seems to be in principle just on the polygons border, and it's considered `OUT` just because of floating point precision.
For this reason I added the second condition to ignore the result if the point is actually not inside any of the polygons of the resulting mesh.

If you have ideas on how to test this better (I do not seem to have found a way to consider tolerance for the `in` function) I believe that would actually be the best approach.

## Benchmarks
I have performed some benchmarks in Julia 1.11-rc2 under windows with the following code:
```julia
using Meshes
using BenchmarkTools
using Random
using StableRNGs

randpoint2(n) = [Point(ntuple(i -> rand(Float64), 2)) for _ in 1:n]

bench(ns) = for n in ns
    Random.seed!(n) # Ensure same points are generated for each n
    pts = randpoint2(n)
    b = @benchmark tesselate($pts, VoronoiTesselation(StableRNG(2024)))
    println("Results for $n points:")
    println()
    display(b)
    println()
end
```

which gives the following results:

### Current version (v0.47.8)
```julia-repl
julia> bench([10, 100, 1000, 10000])
Results for 10 points:

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  197.500 μs …  12.322 ms  ┊ GC (min … max): 0.00% … 96.35%
 Time  (median):     206.600 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   229.758 μs ± 234.335 μs  ┊ GC (mean ± σ):  3.52% ±  3.66%

  ▅█▇▅▄▄▄▄▃▂▂▁▁▁▁▁                                              ▂
  █████████████████████▇▆▆▇▆▆▇▆▆▇▇▇▆▇▆▆▆▅▅▅▆▇▇▅▅▅▆▃▅▅▅▅▆▅▅▅▅▄▄▄ █
  198 μs        Histogram: log(frequency) by time        435 μs <

 Memory estimate: 146.22 KiB, allocs estimate: 1954.

Results for 100 points:

BenchmarkTools.Trial: 2157 samples with 1 evaluation.
 Range (min … max):  2.046 ms …  12.608 ms  ┊ GC (min … max): 0.00% … 79.06%
 Time  (median):     2.131 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.311 ms ± 740.016 μs  ┊ GC (mean ± σ):  4.57% ±  9.95%

  █▇▅▄▃▂▁▁▁
  ██████████▇▇▇▆▆▅▃▅▃▅▅▅▅▅▅▁▅▃▁▃▁▁▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▄▅▇▇▅▅▅ █
  2.05 ms      Histogram: log(frequency) by time      6.07 ms <

 Memory estimate: 1.20 MiB, allocs estimate: 13821.

Results for 1000 points:

BenchmarkTools.Trial: 187 samples with 1 evaluation.
 Range (min … max):  22.325 ms … 44.705 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     25.416 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.812 ms ±  4.288 ms  ┊ GC (mean ± σ):  3.52% ± 6.40%

  ▇▅▄▃▆█▄█▂        ▂
  █████████▇▇█▃▇▆▆▇█▆▅▇▄▅▄▄▆▅▃▁▃▄▃▁▃▄▃▅▁▃▃▁▁▃▃▃▃▁▁▁▁▁▃▃▁▃▃▁▁▃ ▃
  22.3 ms         Histogram: frequency by time        40.9 ms <

 Memory estimate: 10.52 MiB, allocs estimate: 136009.

Results for 10000 points:

BenchmarkTools.Trial: 3 samples with 1 evaluation.
 Range (min … max):  2.066 s …  2.083 s  ┊ GC (min … max): 0.44% … 0.42%
 Time  (median):     2.069 s             ┊ GC (median):    0.42%
 Time  (mean ± σ):   2.073 s ± 9.602 ms  ┊ GC (mean ± σ):  0.29% ± 0.25%

  █        █                                             █
  █▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  2.07 s        Histogram: frequency by time        2.08 s <

 Memory estimate: 106.51 MiB, allocs estimate: 1360405.
```

### After this PR
```julia-repl
julia> bench([10, 100, 1000, 10000])
Results for 10 points:

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  143.500 μs …  17.926 ms  ┊ GC (min … max): 0.00% … 97.78%
 Time  (median):     150.100 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   167.550 μs ± 236.447 μs  ┊ GC (mean ± σ):  4.01% ±  3.31%

  ▆█▇▅▃▃▄▄▃▂▂▁▁▁                                                ▁
  ███████████████▇█▇▇▇▇▆▇▇▆▆▆▆▆▆▅▅▅▅▄▅▅▄▅▅▄▄▅▅▅▄▅▅▅▅▅▅▅▅▅▅▅▅▅▅▅ █
  144 μs        Histogram: log(frequency) by time        336 μs <

 Memory estimate: 124.20 KiB, allocs estimate: 1476.

Results for 100 points:

BenchmarkTools.Trial: 2602 samples with 1 evaluation.
 Range (min … max):  1.637 ms …  19.856 ms  ┊ GC (min … max): 0.00% … 89.87%
 Time  (median):     1.711 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.918 ms ± 930.405 μs  ┊ GC (mean ± σ):  6.26% ± 10.52%

  █▆▄▃▂▂▁
  █████████▇█▇▆▅▇▁▄▁▃▃▁▁▁▃▃▁▁▁▁▁▁▁▁▃▁▁▁▁▁▁▃▃▅▅▆▃▅▃▁▅▃▁▃▁▃▁▁▃▅ █
  1.64 ms      Histogram: log(frequency) by time      7.05 ms <

 Memory estimate: 1021.17 KiB, allocs estimate: 9469.

Results for 1000 points:

BenchmarkTools.Trial: 242 samples with 1 evaluation.
 Range (min … max):  18.067 ms … 33.481 ms  ┊ GC (min … max): 0.00% … 42.22%
 Time  (median):     19.552 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.603 ms ±  2.643 ms  ┊ GC (mean ± σ):  4.60% ±  8.39%

     ▂▇█▂▅▅▁
  ▆▆▆███████▅▅▄▅▅▅▄▃▃▃▃▃▅▃▄▃▃▃▄▁▁▃▁▁▃▄▁▃▃▃▁▁▃▃▄▄▃▃▁▁▃▃▁▃▁▁▁▃▃ ▃
  18.1 ms         Histogram: frequency by time        29.3 ms <

 Memory estimate: 8.35 MiB, allocs estimate: 87556.

Results for 10000 points:

BenchmarkTools.Trial: 3 samples with 1 evaluation.
 Range (min … max):  2.053 s …   2.115 s  ┊ GC (min … max): 0.00% … 0.51%
 Time  (median):     2.064 s              ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.077 s ± 32.766 ms  ┊ GC (mean ± σ):  0.17% ± 0.30%

  █         █                                             █
  █▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  2.05 s         Histogram: frequency by time        2.11 s <

 Memory estimate: 84.92 MiB, allocs estimate: 862476.
```