@disberd 2024-08-06
Tests now seem to fail due to the re-introduction of the latitude check in CoordRefSystems, on est lines not really by the contents of this PR