@juliohm 2024-08-06
The problem is caused by the fact that the `point in poly` predicate with `Ngon` is not using the more robust algorithm implemented in the general case:

https://github.com/JuliaGeometry/Meshes.jl/blob/430b38f46c31b98388a41e75ed9502ef236edb5b/src/predicates/in.jl#L140-L151

We need to benchmark these two algorithms later in a separate PR.